overflow fixed

+++ DirectXApp/Content/shaders/render.hlsl
@@ -6,5 +6,5 @@ float4 main(PixelShaderInput input) : SV_TARGET
 	uint2 intPos = uint2(round(normalized * float2(resolution)-0.5));
 	uint index = resolution.x * intPos.y + intPos.x;
 
-	return float4(pow(accum[index].color / iterations, 1.0/2.2), 1.0);
+	return float4(pow(accum[index].color / double(iterations), 1.0/2.2), 1.0);
 }
